refactor:Split and add test cases
